requirements.txt
av==14.0.1
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.5.0
faster-whisper==1.1.1
filelock==3.16.1
flatbuffers==24.12.23
fsspec==2024.12.0
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.5
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
openai-whisper==20240930
packaging==24.2
protobuf==5.29.3
PyAudio==0.2.14
pydantic==1.10.20
pyreadline3==3.5.4
PyYAML==6.0.2
readchar==4.2.1
regex==2024.11.6
requests==2.32.3
simpleaudio==1.0.4
six==1.17.0
sympy==1.13.1
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
voicevox_core @ https://github.com/VOICEVOX/voicevox_core/releases/download/0.15.5/voicevox_core-0.15.5+cpu-cp38-abi3-win_amd64.whl#sha256=6ee8ff32ce4fc42b946e000997a3685306910422c97f69d645eba028f4cd3f9a
